build: release 25.8.5

+++ egantt/src/main/java/com/egantt/drawing/painter/range/iterator/BasicRangeIterator.java
@@ -37,7 +37,7 @@ public class BasicRangeIterator extends AbstractRangeIterator
 	   this.lastData = data;
 	   this.data = iter.next();
 
-		long x2 = transform.transform(data, (long) width);
+		long x2 = transform.transform(data, width);
 
 	   int width = (int) x2 - (int) lastX;
 	   if (width <= 0)
